+++ CHANGELOG.md
@@ -3,6 +3,12 @@ selenium-liatrio CHANGELOG
 
 This file is used to list changes made in each version of the selenium-liatrio cookbook.
 
+0.2.3
+-----
+##### Ticket: LTA-120 by Justin Bankes
+- Created Spec for Default. 
+
+
 0.1.0
 -----
 - Drew Holt - Initial release of selenium-liatrio